requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.3
PyMySQL==1.0.3
Flask-JWT-Extended==4.4.4
Flask-Validator==1.2.1
reportlab==3.6.13  # For generating PDF balance sheets
